Don't ignore the first line of a changelog
It seems that, at some point in the past, the first line of a changelog
would cause problems for the script, so it was ignored. The format must
have changed, at least for Ubuntu, and now the script misses all patches
from the latest release.

So stop ignoring the first line. This doesn't seem to break anything for
Ubuntu, and the other distros never worked reliably.