Don't always ignore a commit title's last word
If a commit title is too long, the changelog will split it in two lines;
so, in order to find it through simple grep, it's important to only
consider the first 72 characters or so.

A problem with this is that a long word might start within the 72
character limit and still be moved to the next line. The original
solution was to always remove the last word among the first 72
characters, but that causes many short commit titles to look the same
for the script.

Instead, only remove the last word if the commit title has at least 70
characters.